+++ lua/linny/init.lua
@@ -1,0 +1,5 @@
+local M = {}
+
+M.version = require('linny.version')
+
+return M